start writing docs for observable plot components (#342)
* start writing docs for observable plot components

* more work on chart component docs
